+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## 1.1.0 - 2021-xx-xx
 
 ### Added
-- Added `config.meter.seedMockData` which allows for mock meter(s) to be used
+- Added `config.meter.addMockMeters` which allows for mock meter(s) to be used
   for development.
 
 ## 1.0.0 - 2021-07-22